// src/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
  JSON: any;
  Upload: any;
};
// ...
export type Introduction = {
  __typename?: 'Introduction';
  avatar: UploadFileEntityResponse;
  createdAt?: Maybe<Scalars['DateTime']>;
  description: Scalars['String'];
  fullName: Scalars['String'];
  publishedAt?: Maybe<Scalars['DateTime']>;
  updatedAt?: Maybe<Scalars['DateTime']>;
};

export type IntroductionEntity = {
  __typename?: 'IntroductionEntity';
  attributes?: Maybe<Introduction>;
  id?: Maybe<Scalars['ID']>;
};

export type IntroductionEntityResponse = {
  __typename?: 'IntroductionEntityResponse';
  data?: Maybe<IntroductionEntity>;
};
// ...
export type UploadFileEntity = {
  __typename?: 'UploadFileEntity';
  attributes?: Maybe<UploadFile>;
  id?: Maybe<Scalars['ID']>;
};

export type UploadFileEntityResponse = {
  __typename?: 'UploadFileEntityResponse';
  data?: Maybe<UploadFileEntity>;
};